add change message to push.sh

+++ push.sh
@@ -1,7 +1,15 @@
+change=$1
+
+if [ $# -eq 0 ]
+then
+    echo "Please specify a change"
+    exit 1
+fi
+
 rm .vimrc
 cp ~/.vimrc .
 
 git add --all
-git commit -m "update vimrc"
+git commit -m "$change"
 git push -u origin master
